requirements.txt
requests
pandas
nltk
textblob
python-dotenv
numpy
transformers
torch
datasets
tokenizers
matplotlib
seaborn
flask